fix: major refactor
